refactor: изменение расположения контрактов в директории libs
