+++ project/libs/contracts/.eslintrc.json
