+++ project/libs/contracts/tsconfig.json
