+++ project/tsconfig.base.json
@@ -16,7 +16,7 @@
     "baseUrl": ".",
     "paths": {
       "@project/services": ["libs/services/src/index.ts"],
-      "@project/shared/contracts": ["libs/shared/contracts/src/index.ts"],
+      "@project/contracts": ["libs/contracts/src/index.ts"],
       "@project/utils/utils-core": ["libs/utils/utils-core/src/index.ts"]
     }
   },
